requirements.txt
aiogram>=3.7,<4.0
python-dotenv>=1.0,<2.0
openpyxl>=3.1,<4.0
